+++ src/main/kotlin/us/frollo/frollosdk/database/SDKDatabase.kt
@@ -102,7 +102,7 @@ import us.frollo.frollosdk.model.coredata.user.User
         Address::class,
         ServiceOutage::class
     ],
-    version = 18, exportSchema = true
+    version = 19, exportSchema = true
 )
 
 @TypeConverters(Converters::class)
@@ -176,7 +176,8 @@ abstract class SDKDatabase : RoomDatabase() {
                     MIGRATION_14_15,
                     MIGRATION_15_16,
                     MIGRATION_16_17,
-                    MIGRATION_17_18
+                    MIGRATION_17_18,
+                    MIGRATION_18_19
                 )
                 .build()
         }
@@ -756,6 +757,20 @@ abstract class SDKDatabase : RoomDatabase() {
                 // START - Add column sharing_expires_at
                 database.execSQL("ALTER TABLE `consent` ADD COLUMN `sharing_expires_at` TEXT")
                 // END - Add column sharing_expires_at
+            }
+        }
+
+        private val MIGRATION_18_19: Migration = object : Migration(18, 19) {
+            override fun migrate(database: SupportSQLiteDatabase) {
+
+                // WARNING: DO NOT USE "BEGIN TRANSACTION" & "COMMIT" as on latest room version
+                // looks like it does it by default. If we add these we will see the error stated in
+                // https://frollo.atlassian.net/browse/WA-3067
+
+                // New changes in this migration:
+                // 1) Alter table - transaction_model - Add columns merchant_image_url
+
+                database.execSQL("ALTER TABLE `transaction_model` ADD COLUMN `merchant_image_url` TEXT")
             }
         }
     }